+++ kernel.py
@@ -3,9 +3,9 @@ import numpy as np
 
 kernel_lib = cdll.LoadLibrary('./libkernel.so')
 # void build_kernel(
-#         double *data,
-#         const void *indices,
-#         const void *indptr,
+#         float *data,
+#         const int32_t *indices,
+#         const int64_t *indptr,
 #         size_t rows,
 #         const double *train_x,
 #         const double *x,
@@ -15,8 +15,8 @@ kernel_lib = cdll.LoadLibrary('./libkernel.so')
 #         double power_dir)
 kernel_lib.build_kernel.restype = None
 kernel_lib.build_kernel.argtypes = [
-        POINTER(c_double),
-        POINTER(c_int64),
+        POINTER(c_float),
+        POINTER(c_int32),
         POINTER(c_int64),
         c_size_t,
         POINTER(c_double),
@@ -29,15 +29,15 @@ kernel_lib.build_kernel.argtypes = [
 
 
 def build_kernel(graph, train_x, x, l_pts, l_dir, gamma_pts, gamma_dir):
-        if graph.data.dtype != np.float64 \
-          or graph.indices.dtype != np.int64 \
+        if graph.data.dtype != np.float32 \
+          or graph.indices.dtype != np.int32 \
           or graph.indices.dtype != graph.indices.dtype \
           or train_x.dtype != np.float64 \
           or x.dtype != np.float64:
                 raise TypeError("miss-matched type")
 
-        data_p = graph.data.ctypes.data_as(POINTER(c_double))
-        indices_p = graph.indices.ctypes.data_as(POINTER(c_int64))
+        data_p = graph.data.ctypes.data_as(POINTER(c_float))
+        indices_p = graph.indices.ctypes.data_as(POINTER(c_int32))
         indptr_p = graph.indptr.ctypes.data_as(POINTER(c_int64))
         trainx_p = train_x.ctypes.data_as(POINTER(c_double))
         x_p = x.ctypes.data_as(POINTER(c_double))
